refactor(console): split importer engines into focused modules
