+++ console/shared/package.json
@@ -39,12 +39,12 @@
     "./styles/app.css": "./styles/app.css",
     "./components/*": "./components/*",
     "./svelte-config": "./svelte-config.js",
-    "./importer/moobot": "./lib/importer/moobot.ts",
+    "./importer/moobot": "./lib/importer/moobot/index.ts",
     "./importer/caps": "./lib/importer/caps.ts",
     "./importer/types": "./lib/importer/types.ts",
     "./importer/validate": "./lib/importer/validate.ts",
     "./importer/streamelements": "./lib/importer/streamelements.ts",
-    "./importer/streamlabs-desktop": "./lib/importer/streamlabsdesktop.ts"
+    "./importer/streamlabs-desktop": "./lib/importer/streamlabs-desktop/index.ts"
   },
   "dependencies": {
     "@luzir/bolota": "0.1.6",
